Fix issues with user procedures
The eval_sequence method was broken, and our treatment of the
environment for procedures was incorrect. For every invocation of a
procedure, we need to clone the top-most frame which contains local
variables. This allows recursion to work, otherwise arguments are
changed in the middle of execution. E.g. the factorial procedure always
returned 1 because eventually the base case is reached overriding
previous values of n.