Fix some bugs with recursive functions
In the VM I forgot that when we do a SAVE we simply push to the end of
the stack. After procedure calls I was resetting SP but I was not
resetting the actual stack.

In the compiler there were some issues with registers not being saved
properly and on restore variables weren't being marked as in a register.